requirements.txt
torch==2.5.1
numpy==1.26.4
matplotlib==3.9.0
tensorboard==2.18.0
argparse
tqdm==4.67.0
pokers-db==2.2
python-dotenv==1.0.1
requests==2.32.3
seaborn==0.13.2
pandas==2.2.2
PyQt5==5.15.11
PyYAML==6.0.1